README: mention the Xorg.wrap bypass
Xorg.wrap default system of autodetection of root rights requirement is
wrong in every case I've seen it used.  The result is the Xorg server
runs as the root user when it is not necessary.

The main autodetection failure case is when the proprietary nvidia
drivers are used, incorrectly assuming Xorg needs to be run as root.