sx: test -a may not be short-circuiting
I was under the false assumption that -a, like &&, was used
short-circuit semantics.  However the behaviour of -a is not specified
by POSIX.

At least on Linux using bash or dash the behaviour is not
short-circuiting.

As there's no reason to test the RHS if the LHS is false I've switched
back to [ ] && notation which is defined to be short-circuiting.