sx: no longer attempt to send SIGKILL
Sending SIGKILL is not an appropriate response in most cases especially
after an arbitrary timeout.  If signals are being ignored it's likely
due to D state.  If the process is busy spinning in an infinite loop
then bug reports need to be made.  Papering over significant issues is
worse for everyone in the end.

This commit also trusts the PID we get is an xorg server process.  I
have no real reason to check outside of paranoia which borders on theater.

Instead I check the PID is still alive and send TERM.  Then we wait
until it is done.  Anything else is surely abnormal and needs manual
intervention.