sx: traps can explicitly set an exit status
The prior commit 666b87e would still fail if the sx program itself was
killed leaving an xorg server behind.  While trap handlers aren't
typically allowed to change the exit status, if we explicitly exit with
a value then it is honoured.

This works around the subshell setting an exit status of 138 (due to
SIGUSR1) by explicitly using the return value from the last wait or the
last command if the server never started at all.