sx: correct the usage of ps
There appears to be no indication according to the POSIX synopsis[1] of
the ps command that it shall take a PID argument as positional.  The
rest of the document doesn't suggest such a case either.

What the document does provide is the `-p' flag to specify a list of
PIDs with which ps will narrow the results.

1. https://pubs.opengroup.org/onlinepubs/9699919799/utilities/ps.html