doc: Further rationale
Note that startx itself does call upon admins to improve the script for
their particular situation.

This also removes the example using DESTDIR as any package maintainer
would already be familiar with this standard and expect it to work
implicitly.  For those who aren't familiar it merely serves as minor
confusion.