sx: The signal saga concludes
This ought to be the last effort on signal handling.

Prior to this design the resending of the signals was ineffective due to
the cleanup function explicitly calling exit.  As a result this would
prevent the subsequent commands in the trap from executing.

Instead a variable is employed to guard the exit on condition of not
receiving an INT signal.  On UNIX systems this particular signal needs
to be resent such that the calling program knows the process was
interrupted.  See https://www.cons.org/cracauer/sigint.html

This situation exists because, besides bash, all shells treat EXIT traps
as a mechanism to execute code after the program ends, as if the code it
runs was pasted at the end of the file, rather than to run it however the
program terminated.