sx: several minor reworks
Calling exit directly in the relevant trap instead of using an argument
to decide when to trigger an exit in the cleanup function.

Additionally the EXIT trap is ignored in the INT trap so that when
running under bash the EXIT trap does not call cleanup again.

As `xauth remove :n' does not trigger warnings or problems for nonextant
entries it is safe to set the cleanup traps prior to using `xauth add'.
This removes a very minor potential race condition between adding an
entry and the script being interrupted before cleanup traps have been
established.  Suggested by @tomty89

Finally, attempt to make the main comment a little less awkward.