sx: hopefully the final chapter in signal handling
Bash implements EXIT traps in such a manner that it is always invoked
when the shell exits for any reason, such as interruption, termination
or simply coming to a natural end.

As far as I'm aware there are no other shells which implement EXIT traps
in such a manner and are only evaluated when the shell exits normally.

As a consequence each relevant signal has to be handled, and to prevent
double execution when sh happens to be bash, the EXIT trap must be reset
while doing so.

In addition I would also like sx to maintain the WIFSIGNALED status by
re-signalling itself with the signal it received before the cleanup
function was run.  This keeps the parent informed about what signal the
child ended with.

Because of this change sx will no longer always exit with the exit
status of Xorg and so makes it a breaking change.  sx will still attempt
to exit with the exit status of Xorg under non-signalled exit conditions.