Leave xauth default -- in the HOME
Ugly as it is, it's up to X to decide where to put this. I found this
issue when trying to use firejail. I suspect breakages might be possible
when trying to forward X over a socket too. To avoid this kind of breakage
I suggest leaving it at "$HOME/.Xauthority".

It is nice to put Xauthority in the data dir where it really should go,
but it predates that standard.